KAN-125 <message: unscan된 파일 리스트 가져오기>
