+++ src/main/java/com/GASB/anti_malware/controller/FileController.java
@@ -56,14 +56,4 @@ public class FileController {
             return ResponseDto.ofFail(e.getMessage());
         }
     }
-
-    @GetMapping("/unscan")
-    public ResponseDto<List<FileListDto>> getUncheckedFiles() {
-        try {
-            List<FileListDto> response = fileStatusService.getFileWithUnScannedVt();
-            return ResponseDto.ofSuccess(response);
-        } catch(Exception e){
-            return ResponseDto.ofFail(e.getMessage());
-        }
-    }
 }
