+++ src/main/java/com/GASB/anti_malware/controller/VirusTotalController.java
@@ -1,13 +1,14 @@
 package com.GASB.anti_malware.controller;
 
 import com.GASB.anti_malware.model.dto.request.VtRequestDto;
+import com.GASB.anti_malware.model.dto.response.FileListDto;
 import com.GASB.anti_malware.model.dto.response.ResponseDto;
-import com.GASB.anti_malware.model.dto.response.VtReportDto;
 import com.GASB.anti_malware.service.*;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.HashMap;
+import java.util.List;
 import java.util.Map;
 
 @RestController
@@ -18,14 +19,17 @@ public class VirusTotalController {
     private final VtReportSaveService vtReportSaveService;
     private final VtUploadService vtUploadService;
     private final VtAnalysisService vtAnalysisService;
+    private final FileStatusService fileStatusService;
 
     @Autowired
     public VirusTotalController(StoredFileService storedFileService,
-                                VtReportSaveService vtReportSaveService, VtUploadService vtUploadService, VtAnalysisService vtAnalysisService) {
+                                VtReportSaveService vtReportSaveService, VtUploadService vtUploadService,
+                                VtAnalysisService vtAnalysisService, FileStatusService fileStatusService) {
         this.storedFileService = storedFileService;
         this.vtReportSaveService = vtReportSaveService;
         this.vtUploadService = vtUploadService;
         this.vtAnalysisService = vtAnalysisService;
+        this.fileStatusService = fileStatusService;
     }
 
     @PostMapping("/report/{file_id}")
@@ -59,6 +63,16 @@ public class VirusTotalController {
             }
             return ResponseDto.ofSuccess(results);
         } catch (Exception e) {
+            return ResponseDto.ofFail(e.getMessage());
+        }
+    }
+
+    @GetMapping("/unscan")
+    public ResponseDto<List<FileListDto>> getUncheckedFiles() {
+        try {
+            List<FileListDto> response = fileStatusService.getFileWithUnScannedVt();
+            return ResponseDto.ofSuccess(response);
+        } catch(Exception e){
             return ResponseDto.ofFail(e.getMessage());
         }
     }
